+++ Postgres.Marula.Tests/Postgres.Marula.Tests.csproj
@@ -7,13 +7,21 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="NUnit" Version="3.12.0"/>
-		<PackageReference Include="NUnit3TestAdapter" Version="3.16.1"/>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0"/>
+		<PackageReference Include="NUnit" Version="3.12.0" />
+		<PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\Postgres.Marula.DatabaseAccess\Postgres.Marula.DatabaseAccess.csproj"/>
+		<ProjectReference Include="..\Postgres.Marula.DatabaseAccess\Postgres.Marula.DatabaseAccess.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Content Include="appsettings.json">
+			<ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+			<CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
+		</Content>
 	</ItemGroup>
 
 </Project>
