+++ Postgres.Marula.Tests/appsettings.json
@@ -1,0 +1,10 @@
+{
+	"Logging": {
+		"LogLevel": {
+			"Default": "Information"
+		}
+	},
+	"ConnectionStrings": {
+		"Default": "server=localhost; port=5432; database=postgres; username=postgres; password=postgres;"
+	}
+}
